backend/requirements.txt
fastapi
python-multipart
openai
PyMuPDF
langchain
uvicorn
python-docx
Pillow
python-dotenv
